backend/requirements.txt
flask
flask-cors
transformers
torch
scikit-learn
requests
beautifulsoup4
spacy
google-api-python-client
python-dotenv
sentence-transformers
SPARQLWrapper
tldextract